Document OpenID Provider configuration.
The objective of this PR is to document how Opendatasoft supports the OpenID Connect protocol and how a domain administrator can configure the connection to an OpenID Connect provider.

To do so, a reorganization of the domain security configuration section was necessary.

I also used this opportunity to try to improve / simplify the SAML configuration documentation.

Below are the main changes added to this PR:
- Added a new OIDC configuration section.
- Move some documentation content out of the SAML configuration documentation: local / linked user accounts, default login page.
- Removed from SAML documentation references to "account mapper" and "attributes mapper" which IMHO were useless.
- Get rid of IdP acronym and use Identity Provider instead.
- Get rid of the "users federation terminology". Simplified to some extent by replaceing by "single sign-on", which should be what users are looking for.

[ch24112]